Remove use of HTML imports for FF compatibility.
Firefox does not support HTML imports due to the contentions over the final
specification of them. In order to make it work I now do a more typical Webpack
way of including HTML (directly into a string, then set innerHTML). It was a
fairly easy change so should be easily reverted if they become standard.

Also cleaned up code in Hyperlinker, taking advantage of the fact that a Range
object's start and end are such that start is always before end, unlike
Selection objects.